+++ CHANGES.md
@@ -1,3 +1,13 @@
+## 2021-05-23
+
+### Changes
+  * added testcases from https://github.com/eu-digital-green-certificates/dgc-testdata
+  * fixed testcases issues with schema test files from https://github.com/eu-digital-green-certificates/ehn-dgc-schema
+
+### Issues
+  * Some failing unit tests from dgc-testdata, but in my mind those are issues within the provided testdata.
+
+
 ## 2021-05-09
 
 ### Changes